2021.01.10

+++ database/migrations/2020_06_14_225014_create_s_a_p_acc_tables_table.php
@@ -36,7 +36,8 @@ class CreateSAPAccTablesTable extends Migration {
 			$table->string('full_barcode')->nullable();
 
 			$table->string('location')->nullable();
-			$table->integer('qty');
+			// $table->integer('qty');
+			$table->float('qty');
 
 			$table->string('printer');
 			$table->string('printed');
